fix(Aluno): Changing id type from aluno to Long

+++ src/main/java/school/sptech/back_end_PI/controller/AlunoController.java
@@ -47,7 +47,7 @@ public class AlunoController {
     }
 
     @GetMapping("/turma/{id}")
-    public ResponseEntity<List<AlunoResponse>> getByTurmaId(@PathVariable Integer id){
+    public ResponseEntity<List<AlunoResponse>> getByTurmaId(@PathVariable Long id){
         List<Aluno> alunos = service.getByTurmaId(id);
         return ResponseEntity.ok(AlunoMapper.toResponseList(alunos));
     }
